@@ -10,10 +10,15 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "axios": "^1.10.0",
+    "@heroicons/react": "^2.2.0",
+    "@material-tailwind/react": "^2.1.10",
+    "axios": "^1.11.0",
     "bootstrap": "^3.4.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "dotenv": "^17.2.0",
+    "leaflet": "^1.9.4",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-leaflet": "^3.2.5",
     "react-router-dom": "^5.3.4"
   },
   "devDependencies": {
@@ -22,10 +27,13 @@
     "@types/react": "^17.0.80",
     "@types/react-dom": "^17.0.25",
     "@vitejs/plugin-react": "^4.6.0",
+    "autoprefixer": "^10.4.21",
     "eslint": "^8.57.0",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.20",
     "globals": "^13.24.0",
+    "postcss": "^8.5.6",
+    "tailwindcss": "^3.4.17",
     "vite": "^5.4.10"
   },
   "resolve": {
