@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 const api = axios.create({
-  baseURL: 'http://localhost:8000',
+  baseURL: 'http://192.168.1.219:8000',
   headers: {
     'Content-Type': 'application/json',
   }
