@@ -1,11 +1,10 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from .views import BarrioViewSet, PosteViewSet, ClienteViewSet
+from . import views
 
 router = DefaultRouter()
-router.register(r'barrios', BarrioViewSet)
-router.register(r'postes', PosteViewSet)
-router.register(r'clientes', ClienteViewSet)
+router.register(r'clientes', views.ClienteViewSet)
+router.register(r'tipos-servicio', views.TipoServicioViewSet)
 
 urlpatterns = [
     path('', include(router.urls)),
