@@ -30,7 +30,8 @@ urlpatterns = [
     path('api/usuarios/', include('usuarios.urls')),
     path('api/almacenes/', include('almacenes.urls')),
     path('api/contratos/', include('contratos.urls')),
-    path('api/', include('maps.urls')),
+    path('api/maps/', include('maps.urls')),
+    
 
 
     
